add demo file.

+++ sources/easy-ORM-demo/src/main/java/com/jungle/easyorm/demo/data/Student.java
@@ -30,5 +30,5 @@ public class Student extends BaseEntity {
 
     public String mName;
     public int mAge;
-    public double mScore;
+    public float mScore;
 }
